ADD componente mapa e organizando o cod
